==> College_Graduate_Earnings_Prediction/project_files/3 - Final Data Cleaning.ipynb <==
import pickle
import numpy as np
import pandas as pd

X = data = pickle.load(open('merged_earnings_df.pickle', 'rb'))
X

categorical_keys = [
    "STABBR",
    "ACCREDAGENCY",
    "PREDDEG",
    "HIGHDEG",
    "CONTROL",
    "REGION",
    "LOCALE",
    "CCBASIC",
    "CCUGPROF",
    "CCSIZSET",
    "RELAFFIL",
   # "ICLEVEL",
]

X = pd.get_dummies(
    X,
    dummy_na=True,
    columns=[*[col for col in categorical_keys if col in X.columns]],
    drop_first=True,
)

list(X.columns)

X.loc[:, "MD_EARN_AVG"].isna().value_counts()

y = X.MD_EARN_AVG
X.drop("MD_EARN_AVG", axis=1, inplace=True)

X.YEAR.value_counts()

with open("clean_data.pickle", "wb") as f:
    pickle.dump(
        {
            "X": X,
            'y': y
        },
        f,
    )

with open('clean_data.pickle','rb') as f:
    d = pickle.load(f)
    X = d['X']
    y = d['y']
    del(d)

X.isna()

non_empty_cols = (X.isna().sum() / X.shape[0]) != 1
non_empty_col_names = [col for col, v in non_empty_cols.to_dict().items() if v]
X = X[non_empty_col_names]
col_pct = (X.isna().sum() / X.shape[0])
col_pct_nonzero = pd.Series({c: p for c, p in col_pct.to_dict().items() if p > 0})
col_pct_nonzero.sort_values(ascending=False)

naive_impute_avg = [
    "PFTFAC",
    "INEXPFTE",
    "AVGFACSAL",
    "GRAD_DEBT_MDN",
    "C150_4",
    "ADM_RATE",
]
for col in naive_impute_avg:
    X.loc[:,f"{col}_ISNA"] = X[col].isna().astype(int)
    X.loc[:,col] = X[col].fillna(X[col].mean())

X.loc[:,"TUITION_ISNA"] = X["TUITIONFEE_OUT"].isna().astype(int)

X.loc[:, "TUITIONFEE_IN"] = X["TUITIONFEE_IN"].fillna(value=X["TUITIONFEE_PROG"])
X.loc[:, "TUITIONFEE_OUT"] = X["TUITIONFEE_OUT"].fillna(value=X["TUITIONFEE_PROG"])
X.loc[:, "TUITION_OUT_ISNT_IN"] = (X["TUITIONFEE_IN"] != X["TUITIONFEE_OUT"]).astype(
    int
)
X = X.dropna(subset=["TUITIONFEE_OUT"], how="any", axis="rows")
X = X.drop("TUITIONFEE_PROG", axis=1)
X = X.drop("PRGMOFR", axis=1)
y = y[X.index]

X.shape

non_empty_cols = (X.isna().sum() / X.shape[0]) != 1
non_empty_col_names = [col for col, v in non_empty_cols.to_dict().items() if v]
X = X[non_empty_col_names]
col_pct = (X.isna().sum() / X.shape[0])
col_pct_nonzero = pd.Series({c: p for c, p in col_pct.to_dict().items() if p > 0})
col_pct_nonzero.sort_values(ascending=False)

X['MAIN'] = X.MAIN.replace({'Main campus': 1, 'Not main campus': 0})

X.drop(col_pct_nonzero.index, axis = 1, inplace = True)

#drop those entries without any income
X = X[y!=0]
y = y[y!=0]

#then reset the indices
X.reset_index(drop = True, inplace = True)
y.reset_index(drop = True, inplace = True)

X

nonzero_var_cols = []
for col in X.columns:
    if min(X[col]) == max(X[col]):
        print(col)
        print(min(X[col]))
    else:
        nonzero_var_cols.append(col)

X = X[nonzero_var_cols]

with open("clean_notna_data.pickle", "wb") as f:
    pickle.dump(
        {
            "X": X,
            'y': y
        },
        f,
    )
